+++ tests/unit/mcap/test_encoding_fix.py
