+++ tests/unit/mcap/test_recursive_lookup.py
