+++ tests/unit/zipdcm/resources/dcms/x/1-1.dcm
